Make device handling more generic
Code now takes in a single argument indicating the vid:pid of the tablet
as output by lsusb. It also automatically tries to find the correct
endpoints to send the init packet to and receive the response.

Assumes that the interface with 2 endpoints is correct, and that one
will be the in endpoint and the other will be the out endpoint. Also
assumes 1 altsetting per interface.